refactor: move follows module under users module
