+++ src/users/follows/follows.resolver.ts
@@ -1,6 +1,6 @@
 import { Args, Mutation, Resolver } from "@nestjs/graphql";
-import { CurrentUser } from "../auth/decorators/current-user.decorator";
-import { User } from "../users/models/user.model";
+import { CurrentUser } from "../../auth/decorators/current-user.decorator";
+import { User } from "../models/user.model";
 import { FollowsService } from "./follows.service";
 import { CreateFollowInput } from "./models/create-follow.input";
 import { CreateFollowPayload } from "./models/create-follow.payload";
