+++ src/users/follows/models/follow.model.ts
@@ -9,7 +9,7 @@ import {
   PrimaryGeneratedColumn,
   UpdateDateColumn,
 } from "typeorm";
-import { User } from "../../users/models/user.model";
+import { User } from "../../models/user.model";
 
 @ObjectType()
 @Entity()
