+++ src/users/models/user.model.ts
@@ -11,7 +11,7 @@ import {
 } from "typeorm";
 import { RefreshToken } from "../../auth/refresh-tokens/models/refresh-token.model";
 import { FeedItem } from "../../common/models/feed-item.union";
-import { Follow } from "../../follows/models/follow.model";
+import { Follow } from "../follows/models/follow.model";
 import { GroupMember } from "../../groups/group-members/models/group-member.model";
 import { Image } from "../../images/models/image.model";
 import { Like } from "../../likes/models/like.model";
