+++ src/users/users.resolver.ts
@@ -11,8 +11,8 @@ import {
 import { CurrentUser } from "../auth/decorators/current-user.decorator";
 import { FeedItem } from "../common/models/feed-item.union";
 import { Dataloaders } from "../dataloader/dataloader.types";
-import { FollowsService } from "../follows/follows.service";
-import { Follow } from "../follows/models/follow.model";
+import { FollowsService } from "./follows/follows.service";
+import { Follow } from "./follows/models/follow.model";
 import { Group } from "../groups/models/group.model";
 import { Image } from "../images/models/image.model";
 import { Post } from "../posts/models/post.model";
